requirements.txt
connexion
gunicorn
uvicorn
google-cloud-aiplatform>=1.38
google-cloud-texttospeech
google-cloud-secret-manager
google-crc32c
